e2e: cpus-present is a system-wide topology parameter, not per node
As per-node parameter it would give a strong hit of wrong semantics:
as if given amount of CPUs were available in each node. That kind of
setup cannot be configured with Qemu. Furthermore, as a system wide
parameter, this allows launching a VM with only few CPUs online
even in the case of 8-socket systems. Otherwise the minimum would
have been 8 in that case, which could easily oversubscribing CPUs
from typical client host systems.